web: Separate configurable state from the individual modules.
This commit implements all configurable state as a separate module in
an effort to simplify the reasoning behind configurable properties in
djehuty.  For an example of unclarity: Should storage locations be
configured as a property of the database, or as a property of the web
server?  In this commit, we avoid this ambiguity by making
configurable properties a module in itself.

Other than a clarification, this commit takes a large step towards
implementing atomic reconfiguration, or hot-reloadable configuration
files.

* src/Makefile.am: Register “config.py” to the build system.
* src/djehuty/web/database.py: Implement configurable properties by
  using the “config” module.
* src/djehuty/web/locks.py: Likewise.
* src/djehuty/web/wsgi.py: Likewise.
* src/djehuty/web/ui.py: Rework the configuration loading to store all
  configurable state in the “config” module/instance.