web: Embed simplified 'zipfly'.
This commit takes a refactored/stripped version of zipfly and embeds
it as a module instead of using it as a dependency.  The main reason
to do so is that most package managers don't have zipfly packaged.  In
total, it counts less than 100 lines of code, which I don't expect
need any maintenance in the near of distant future because the Zip
format is stable.

By embedding this package, setting up a development environment on
Windows becomes doable within MSYS2/MINGW64.  Creating RPMs for
Fedora and RHEL9 becomes possible without seperately packaging
zipfly. On Debian, the situation is improved, but only Pillow is
missing from the packages.

* guix.scm.in: Remove zipfly from the dependency list.
* requirements.txt: Likewise.
* rpmbuild/SPECS/djehuty.spec.in: Likewise.
* src/djehuty/web/wsgi.py: Import the embedded zipfly.
* src/djehuty/web/zipfly.py: New file.