web: database: Implement stricter security practices for RDFLib.
This commit implements an audit hook to detect using RDFLib for
anything other than connecting to the configured SPARQL endpoint.
It's not entirely comprehensive, because sending a SPARQL query that
will attempt to LOAD another resource isn't detected by this audit
hook because it's the SPARQL endpoint that performs the request, not
RDFLib.  Therefore, this commit also adds a clause to prevent
executing LOAD queries on the SPARQL endpoint.

* src/djehuty/web/database.py: Implement and install
  ‘rdflib_network_audit_hook’.