web: html_templates: dataset: Implement visual indicator for “download all”.
On a slow day, it can take a while (in the order of minutes) before
the ZIP file download starts.  This is due to having to check whether
the files are available on S3, and streaming them into the ZIP file.

Therefore, this commit implements a pop-up message that gives a visual
indicator that clicking the “download all files” button has been received.

* src/djehuty/web/resources/html_templates/dataset.html: Implement div
  to show the message in.
* src/djehuty/web/resources/static/css/pub.css: Implement styling bits.
* src/djehuty/web/resources/static/js/dataset_landing_page.js:
  Implement event handling to display the message.