web: database: Avoid wrongfully copying “derived_from” in new versions.
When there is no value for “derived_from”, the code before this commit
would generate a “Derived From” custom field containing the empty
list.  That empty list would in turn show in the edit dataset form.

This commit prevents generating the the “derived_from” field
containing the empty list.  It also cleans up some code found along
the way of tracing this bug.

* src/djehuty/web/database.py: Avoid appending a non-existent “Derived
  From” custom field.