web: Limit ability to input HTML to a list of known tags.
This commit implements mitigations to a Cross-Site-Scripting
vulnerability found by Aaron Liebig.  Thank you very much for the
detailed report and help in finding a solution.  In a follow-up commit the
Content-Security-Policy will be tightened to not allow 'unsafe-inline'
scripting and styling to further mitigate the effect of injecting
JavaScript into pages.

* src/djehuty/utils/constants.py: Define the “allowed_html_tags” constant.
* src/djehuty/utils/convenience.py: Implement ‘contains_disallowed_html’ and
  ‘encode_html’.
* src/djehuty/web/resources/static/css/form.css: Remove the link
  button in the WYSIWYG editor.
* src/djehuty/web/validator.py: Implement “error_on_disallowed_html”
  parameter in ‘string_value’; encode all HTML except for explicitely
  allowed tags.

Co-authored-by: Roel Janssen <r.r.e.janssen@tudelft.nl>